created new server file with code from doc
